requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.1.0
coverage==6.4.1
dj-database-url==0.5.0
Django==4.0.5
django-filter==21.1
django-phonenumber-field==6.1.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
phonenumbers==8.12.50
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
stripe==3.5.0
tomli==2.0.1
urllib3==1.26.9
whitenoise==6.2.0
